+++ common/RELEASES.md
@@ -691,3 +691,13 @@ Based on:
 - [typescript v0.38.0] common
 ### Releases
 - [NPM v0.38.0] https://www.npmjs.com/package/@codat/common/v/0.38.0 - common
+
+## 2023-09-01 00:03:47
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.77.0 (2.91.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.39.0] common
+### Releases
+- [NPM v0.39.0] https://www.npmjs.com/package/@codat/common/v/0.39.0 - common